backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
motor>=3.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0
slowapi>=0.1.9
python-multipart>=0.0.6
